requirements.txt
discord
python-dotenv
asyncio
pillow
openai